Database: Migrate -- Change type of column price in data

+++ database/migrations/2021_10_28_055229_create_bills_table.php
@@ -17,7 +17,7 @@ class CreateBillsTable extends Migration
             $table->id();
             $table->bigInteger('user_id')->unsigned()->nullable();
             $table->foreign('user_id')->references('id')->on('users')->onUpdate('cascade')->onDelete('cascade');
-            $table->float('total_price');
+            $table->decimal('total_price',12,2);
             $table->timestamps();
         });
     }
